// frontend/src/app/models/notification/notification-type.enum.ts
export enum NotificationType {
  FRIENDSHIP_REQUEST = 'FRIENDSHIP_REQUEST',
  FRIENDSHIP_ACCEPTED = 'FRIENDSHIP_ACCEPTED',
  MEMBERSHIP_REQUEST = 'MEMBERSHIP_REQUEST',
  MEMBERSHIP_ACCEPTED = 'MEMBERSHIP_ACCEPTED',
  MEETING_SCHEDULED = 'MEETING_SCHEDULED',
  MEETING_STARTED = 'MEETING_STARTED',
  PEER_MESSAGE = 'PEER_MESSAGE',
  ROOM_MESSAGE = 'ROOM_MESSAGE',
  WARNING = 'WARNING',
  ERROR = 'ERROR',
  TEST = 'TEST'
}